Update Test for Take into a separate file
