+++ fixtures/data.json
@@ -5,5 +5,10 @@
     { "name": "topo", "age": 18 },
     { "name": "pedro", "age": 19 }
   ],
-  "numbers": [1, 2, 3, 4, 5]
+  "numbers": [1, 2, 3, 4, 5],
+  "children": [
+        { "name": "ana", "sex": "f" },
+        { "name": "fosto", "sex": "m" },
+        { "name": "jane", "sex": "f" }
+  ]
 }
